third_party/blink/web_tests/platform/generic/fast/css/default-bidi-css-rules-expected.txt

This test checks the default rules for direction and unicode-bidi CSS properties.

PASS styleOf("div", {}).direction is "ltr"
FAIL styleOf("div", {}).unicodeBidi should be -webkit-isolate. Was normal.
PASS styleOf("div", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("div", {"dir":"ltr"}).unicodeBidi is "isolate"
PASS styleOf("div", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("div", {"dir":"rtl"}).unicodeBidi is "isolate"
PASS styleOf("div", {"dir":"auto"}).direction is "ltr"
PASS styleOf("div", {"dir":"auto"}).unicodeBidi is "isolate"
PASS styleOf("div", {"dir":""}).direction is "ltr"
PASS styleOf("div", {"dir":""}).unicodeBidi is "isolate"
PASS styleOf("span", {}).direction is "ltr"
PASS styleOf("span", {}).unicodeBidi is "normal"
PASS styleOf("span", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("span", {"dir":"ltr"}).unicodeBidi is "isolate"
PASS styleOf("span", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("span", {"dir":"rtl"}).unicodeBidi is "isolate"
PASS styleOf("span", {"dir":"auto"}).direction is "ltr"
PASS styleOf("span", {"dir":"auto"}).unicodeBidi is "isolate"
PASS styleOf("span", {"dir":""}).direction is "ltr"
PASS styleOf("span", {"dir":""}).unicodeBidi is "isolate"
PASS styleOf("bdi", {}).direction is "ltr"
PASS styleOf("bdi", {}).unicodeBidi is "isolate"
PASS styleOf("bdi", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("bdi", {"dir":"ltr"}).unicodeBidi is "isolate"
PASS styleOf("bdi", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("bdi", {"dir":"rtl"}).unicodeBidi is "isolate"
PASS styleOf("bdi", {"dir":"auto"}).direction is "ltr"
PASS styleOf("bdi", {"dir":"auto"}).unicodeBidi is "isolate"
PASS styleOf("bdi", {"dir":""}).direction is "ltr"
PASS styleOf("bdi", {"dir":""}).unicodeBidi is "isolate"
PASS styleOf("output", {}).direction is "ltr"
PASS styleOf("output", {}).unicodeBidi is "isolate"
PASS styleOf("output", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("output", {"dir":"ltr"}).unicodeBidi is "isolate"
PASS styleOf("output", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("output", {"dir":"rtl"}).unicodeBidi is "isolate"
PASS styleOf("output", {"dir":"auto"}).direction is "ltr"
PASS styleOf("output", {"dir":"auto"}).unicodeBidi is "isolate"
PASS styleOf("output", {"dir":""}).direction is "ltr"
PASS styleOf("output", {"dir":""}).unicodeBidi is "isolate"
PASS styleOf("bdo", {}).direction is "ltr"
PASS styleOf("bdo", {}).unicodeBidi is "bidi-override"
PASS styleOf("bdo", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("bdo", {"dir":"ltr"}).unicodeBidi is "bidi-override"
PASS styleOf("bdo", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("bdo", {"dir":"rtl"}).unicodeBidi is "bidi-override"
PASS styleOf("bdo", {"dir":"auto"}).direction is "ltr"
FAIL styleOf("bdo", {"dir":"auto"}).unicodeBidi should be bidi-override isolate. Was isolate.
PASS styleOf("bdo", {"dir":""}).direction is "ltr"
PASS styleOf("bdo", {"dir":""}).unicodeBidi is "bidi-override"
PASS styleOf("textarea", {}).direction is "ltr"
PASS styleOf("textarea", {}).unicodeBidi is "normal"
PASS styleOf("textarea", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("textarea", {"dir":"ltr"}).unicodeBidi is "isolate"
PASS styleOf("textarea", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("textarea", {"dir":"rtl"}).unicodeBidi is "isolate"
PASS styleOf("textarea", {"dir":"auto"}).direction is "ltr"
PASS styleOf("textarea", {"dir":"auto"}).unicodeBidi is "plaintext"
PASS styleOf("textarea", {"dir":""}).direction is "ltr"
PASS styleOf("textarea", {"dir":""}).unicodeBidi is "isolate"
PASS styleOf("pre", {}).direction is "ltr"
PASS styleOf("pre", {}).unicodeBidi is "normal"
PASS styleOf("pre", {"dir":"ltr"}).direction is "ltr"
PASS styleOf("pre", {"dir":"ltr"}).unicodeBidi is "isolate"
PASS styleOf("pre", {"dir":"rtl"}).direction is "rtl"
PASS styleOf("pre", {"dir":"rtl"}).unicodeBidi is "isolate"
PASS styleOf("pre", {"dir":"auto"}).direction is "ltr"
PASS styleOf("pre", {"dir":"auto"}).unicodeBidi is "plaintext"
PASS styleOf("pre", {"dir":""}).direction is "ltr"
PASS styleOf("pre", {"dir":""}).unicodeBidi is "isolate"
PASS successfullyParsed is true

TEST COMPLETE
